root_hash_checker
